requirements.txt
django
djangorestframework
dotenv
django-cors-headers
djangorestframework-simplejwt
djangorestframework-authtoken
drf-spectacular
psycopg2-binary
requests
